requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.52.0
databricks-sdk>=0.30.0
databricks-sql-connector>=3.0.0
pydantic>=2.0.0
